@@ -17,31 +17,31 @@
         "source": "https://github.com/php-tuf/composer-stager"
     },
     "require": {
-        "php": "^7.4 || ^8.0",
+        "php": ">=8.1.0",
         "ext-json": "*",
-        "symfony/filesystem": "^4.4 || ^5.2 || ^6.1",
-        "symfony/process": "^4.4 || ^5.2 || ^6.1"
+        "symfony/filesystem": "^6.2",
+        "symfony/process": "^6.2"
     },
     "require-dev": {
         "dealerdirect/phpcodesniffer-composer-installer": "^1.0",
-        "ergebnis/composer-normalize": "^2.28",
+        "ergebnis/composer-normalize": "^2.30",
         "php-parallel-lint/php-parallel-lint": "^1.3",
         "phpmd/phpmd": "^2.13",
-        "phpro/grumphp-shim": "^1.13",
-        "phpspec/prophecy": "^1.16",
+        "phpro/grumphp-shim": "^1.15",
+        "phpspec/prophecy": "^1.17",
         "phpspec/prophecy-phpunit": "^2.0",
         "phpstan/extension-installer": "^1.2",
-        "phpstan/phpstan": "^1.9",
+        "phpstan/phpstan": "^1.10",
         "phpstan/phpstan-deprecation-rules": "^1.1",
-        "phpstan/phpstan-strict-rules": "^1.4",
-        "phpunit/phpunit": "^9.5",
-        "psalm/phar": "^5.6",
-        "qossmic/deptrac-shim": "^0.24.0",
+        "phpstan/phpstan-strict-rules": "^1.5",
+        "phpunit/phpunit": "^9.6",
+        "psalm/phar": "^5.7",
+        "qossmic/deptrac-shim": "^1.0",
         "slevomat/coding-standard": "^8.8",
         "squizlabs/php_codesniffer": "^3.7",
-        "symfony/config": "^5.4",
-        "symfony/dependency-injection": "^5.4",
-        "symfony/yaml": "^5.4",
+        "symfony/config": "^6.2",
+        "symfony/dependency-injection": "^6.2",
+        "symfony/yaml": "^6.2",
         "thecodingmachine/phpstan-strict-rules": "^1.0"
     },
     "minimum-stability": "dev",
@@ -61,7 +61,7 @@
         "allow-plugins": {
             "dealerdirect/phpcodesniffer-composer-installer": true,
             "ergebnis/composer-normalize": true,
-            "phpro/grumphp-shim": true,
+            "phpro/grumphp-shim": false,
             "phpstan/extension-installer": true
         },
         "optimize-autoloader": true,
