@@ -14,20 +14,11 @@ use PhpTuf\ComposerStager\Infrastructure\Value\PathList\PathList;
 
 final class PhpFileSyncer implements PhpFileSyncerInterface
 {
-    private RecursiveFileFinderInterface $fileFinder;
-
-    private FilesystemInterface $filesystem;
-
-    private PathFactoryInterface $pathFactory;
-
     public function __construct(
-        RecursiveFileFinderInterface $fileFinder,
-        FilesystemInterface $filesystem,
-        PathFactoryInterface $pathFactory
+        private readonly RecursiveFileFinderInterface $fileFinder,
+        private readonly FilesystemInterface $filesystem,
+        private readonly PathFactoryInterface $pathFactory,
     ) {
-        $this->fileFinder = $fileFinder;
-        $this->filesystem = $filesystem;
-        $this->pathFactory = $pathFactory;
     }
 
     public function sync(
@@ -35,7 +26,7 @@ final class PhpFileSyncer implements PhpFileSyncerInterface
         PathInterface $destination,
         ?PathListInterface $exclusions = null,
         ?ProcessOutputCallbackInterface $callback = null,
-        ?int $timeout = ProcessRunnerInterface::DEFAULT_TIMEOUT
+        ?int $timeout = ProcessRunnerInterface::DEFAULT_TIMEOUT,
     ): void {
         set_time_limit((int) $timeout);
 
@@ -85,7 +76,7 @@ final class PhpFileSyncer implements PhpFileSyncerInterface
     private function deleteExtraneousFilesFromDestination(
         PathInterface $destination,
         PathInterface $source,
-        PathListInterface $exclusions
+        PathListInterface $exclusions,
     ): void {
         // There's no reason to look for deletions if the destination is already empty.
         if ($this->destinationIsEmpty($destination)) {
@@ -127,7 +118,7 @@ final class PhpFileSyncer implements PhpFileSyncerInterface
     private function copySourceFilesToDestination(
         PathInterface $source,
         PathInterface $destination,
-        PathListInterface $exclusions
+        PathListInterface $exclusions,
     ): void {
         $sourceFiles = $this->fileFinder->find($source, $exclusions);
 
