@@ -8,8 +8,6 @@ abstract class AbstractPath implements PathInterface
 {
     protected string $cwd;
 
-    protected string $path;
-
     abstract protected function doResolve(string $basePath): string;
 
     /**
@@ -23,10 +21,8 @@ abstract class AbstractPath implements PathInterface
      *   simply assumed to represent a directory, as opposed to a file--even if
      *   it has an extension, which is no guarantee of type.
      */
-    public function __construct(string $path, ?PathInterface $cwd = null)
+    public function __construct(protected readonly string $path, ?PathInterface $cwd = null)
     {
-        $this->path = $path;
-
         // Especially since it accepts relative paths, an immutable path value
         // object should be immune to environmental details like the current
         // working directory. Cache the CWD at time of creation.
